First commit
sdlOOP lib and unit test

this includes a bunch of core files that will be used for the sdlOOP lib
as well as boost unit test project for the lib, as well as a few my own
files for a boost test harness which can be seen in test/TestTracker.h
and a few other files. As time goes on I'll increase the level of
organization and documentation.